Reconfigures settings view
Simplifies UI so only the billing options and log out button are displayed
